requirements.txt
appnope==0.1.0
asgiref==3.2.3
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.1
defusedxml==0.6.0
Django==3.0.2
django-allauth==0.41.0
django-braces==1.14.0
django-filter==2.2.0
django-oauth-toolkit==1.2.0
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
django-rest-multiple-models==2.1.3
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
drf-writable-nested==0.5.4
entrypoints==0.3
idna==2.8
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.2
MarkupSafe==1.1.1
mistune==0.8.4
model-bakery==1.1.0
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
oauthlib==3.1.0
pandocfilters==1.4.2
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.3
ptyprocess==0.6.0
Pygments==2.5.2
PyJWT==1.7.1
pyrsistent==0.15.7
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
pyzmq==18.1.1
requests==2.22.0
requests-oauthlib==1.3.0
Send2Trash==1.5.0
six==1.14.0
social-auth-app-django==3.1.0
-e git+https://github.com/KiyeopYang/social-core.git@497fd7be24413bb0fba11d0b2beebac4ba0eba62#egg=social_auth_core
sqlparse==0.3.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
Unidecode==1.1.1
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
zipp==2.2.0
